refactor: 메서드, 클래스, 패키지 명 변경
