+++ src/test/java/cart/dao/ItemDaoTest.java
@@ -1,10 +1,10 @@
-package cart.repository;
+package cart.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import cart.dao.dto.ItemDto;
 import cart.model.Item;
-import cart.repository.dto.ItemDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
